+++ platforms/fabric-1.15/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinPlayerInteractionManagerForLockedSlotsMovePrevention.java
@@ -4,7 +4,8 @@ import net.minecraft.client.network.ClientPlayerInteractionManager;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.container.SlotActionType;
-import org.anti_ad.mc.ipnext.config.ModSettings;
+import org.anti_ad.mc.ipnext.config.GuiSettings;
+import org.anti_ad.mc.ipnext.event.ContinuousCraftingHandler;
 import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
 import org.anti_ad.mc.ipnext.event.LockedSlotKeeper;
 import org.spongepowered.asm.mixin.Mixin;
@@ -26,12 +27,30 @@ public class MixinPlayerInteractionManagerForLockedSlotsMovePrevention {
                           SlotActionType actionType,
                           PlayerEntity player,
                           CallbackInfoReturnable<ItemStack> cir) {
-        if(ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_QUICK_MOVE_THROW().getValue() &&
-                !LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() &&
-                (actionType == SlotActionType.QUICK_MOVE || (actionType == SlotActionType.THROW && button == 1))) {
-            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, button)) {
+        boolean move = actionType == SlotActionType.QUICK_MOVE;
+        boolean thr = actionType == SlotActionType.THROW;
+        if(!LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() && (move || thr)) {
+            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, thr, button)) {
                 cir.cancel();
             }
+        }
+        if (GuiSettings.INSTANCE.getCONTINUOUS_CRAFTING_SAVED_VALUE().getValue()) {
+            ContinuousCraftingHandler.INSTANCE.setProcessingClick(true);
+        }
+    }
+
+    @Inject(at = @At(value = "TAIL",
+            target = "Lnet/minecraft/client/network/ClientPlayerInteractionManager;clickSlot(IIILnet/minecraft/screen/slot/SlotActionType;Lnet/minecraft/entity/player/PlayerEntity;)V"),
+            method = "clickSlot",
+            cancellable = true)
+    public void postClickSlot(int syncId,
+                          int slotIndex,
+                          int button,
+                          SlotActionType actionType,
+                          PlayerEntity player,
+                          CallbackInfoReturnable<ItemStack> cir) {
+        if (GuiSettings.INSTANCE.getCONTINUOUS_CRAFTING_SAVED_VALUE().getValue()) {
+            ContinuousCraftingHandler.INSTANCE.setProcessingClick(false);
         }
     }
 }