+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/common/mixin/MixinKeyboard.java
@@ -2,6 +2,7 @@ package org.anti_ad.mc.common.mixin;
 
 import net.minecraft.client.Keyboard;
 import net.minecraft.client.gui.screen.Screen;
+import org.anti_ad.mc.common.Log;
 import org.anti_ad.mc.common.input.GlobalInputHandler;
 import org.anti_ad.mc.common.input.GlobalScreenEventListener;
 import org.anti_ad.mc.common.vanilla.Vanilla;
@@ -11,6 +12,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+
 @Mixin(Keyboard.class)
 public class MixinKeyboard {
     @Shadow
@@ -18,7 +20,8 @@ public class MixinKeyboard {
 
     @Inject(method = "onKey", at = @At(value = "TAIL")) // ref: malilib key hook
     private void onKeyLast(long handle, int key, int scanCode, int action, int modifiers, CallbackInfo ci) {
-        if (handle == Vanilla.INSTANCE.window().getHandle()) {
+        if (Vanilla.INSTANCE.mc().isOnThread() && handle == Vanilla.INSTANCE.window().getHandle()) {
+            Log.INSTANCE.debug("TAIL");
             if (Vanilla.INSTANCE.screen() == null) { // non null is handled below
                 GlobalInputHandler.INSTANCE.onKey(key, scanCode, action, modifiers);
             } else {
@@ -32,8 +35,10 @@ public class MixinKeyboard {
             "wrapScreenError(Ljava/lang/Runnable;Ljava/lang/String;Ljava/lang/String;)V"), cancellable = true)
     private void onScreenKey(long handle, int key, int scanCode, int action, int modifiers, CallbackInfo ci) {
         Screen lastScreen = Vanilla.INSTANCE.screen();
+        Log.INSTANCE.debug("INVOKE");
         boolean result = GlobalInputHandler.INSTANCE.onKey(key, scanCode, action, modifiers)
                 || GlobalScreenEventListener.INSTANCE.onKey(key, scanCode, action, modifiers, repeatEvents, true);
+
         if (result || lastScreen != Vanilla.INSTANCE.screen()) { // detect gui change, cancel vanilla
             ci.cancel();
         }