+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinGameRenderer.java
@@ -1,8 +1,6 @@
 package org.anti_ad.mc.ipnext.mixin;
 
 import net.minecraft.client.render.GameRenderer;
-import net.minecraft.client.util.math.MatrixStack;
-import org.anti_ad.mc.ipnext.config.Tweaks;
 import org.anti_ad.mc.ipnext.gui.inject.ScreenEventHandler;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;