+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinMinecraftClient.java
@@ -1,9 +1,14 @@
 package org.anti_ad.mc.ipnext.mixin;
 
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.network.ClientPlayerEntity;
+import net.minecraft.client.option.GameOptions;
 import net.minecraft.client.world.ClientWorld;
-import org.anti_ad.mc.ipnext.config.Tweaks;
+import net.minecraft.entity.player.PlayerInventory;
+import org.anti_ad.mc.ipnext.config.ModSettings;
 import org.anti_ad.mc.ipnext.event.ClientEventHandler;
+import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -16,6 +21,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(MinecraftClient.class)
 public abstract class MixinMinecraftClient {
 
+    @Shadow
+    public ClientPlayerEntity player;
+
+    @Shadow @Final public GameOptions options;
 
     @Inject(at = @At("HEAD"), method = "tick()V")
     public void tick(CallbackInfo info) {
@@ -30,5 +39,20 @@ public abstract class MixinMinecraftClient {
     @Inject(at = @At("RETURN"), method = "joinWorld(Lnet/minecraft/client/world/ClientWorld;)V")
     public void joinWorld(ClientWorld clientWorld, CallbackInfo info) {
         ClientEventHandler.INSTANCE.onJoinWorld(clientWorld);
+    }
+
+    @Inject(at = @At("HEAD"),
+            method = "handleInputEvents()V")
+    public void handleInputEvents(CallbackInfo info) {
+        if(ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_THROW_FOR_NON_STACKABLE().getValue()
+                && PlayerInventory.isValidHotbarIndex(this.player.getInventory().selectedSlot)
+                && (MinecraftClient.getInstance().options.keyDrop.isPressed() || MinecraftClient.getInstance().options.keyDrop.wasPressed())) {
+
+            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(this.player.getInventory().selectedSlot + 36, true, 0)) {
+                IMixinKeyBinding drop = (IMixinKeyBinding) MinecraftClient.getInstance().options.keyDrop;
+                drop.setPressed(false);
+                drop.setTimesPressed(0);
+            }
+        }
     }
 }