+++ platforms/fabric-1.17/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinPlayerInteractionManagerForLockedSlotsMovePrevention.java
@@ -4,7 +4,6 @@ import net.minecraft.client.network.ClientPlayerInteractionManager;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.screen.slot.SlotActionType;
 import org.anti_ad.mc.ipnext.config.GuiSettings;
-import org.anti_ad.mc.ipnext.config.ModSettings;
 import org.anti_ad.mc.ipnext.event.ContinuousCraftingHandler;
 import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
 import org.anti_ad.mc.ipnext.event.LockedSlotKeeper;
@@ -12,8 +11,6 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-import org.anti_ad.mc.common.Log;
 
 @Mixin(ClientPlayerInteractionManager.class)
 public class MixinPlayerInteractionManagerForLockedSlotsMovePrevention {
@@ -29,10 +26,10 @@ public class MixinPlayerInteractionManagerForLockedSlotsMovePrevention {
                           PlayerEntity playerEntity,
                           CallbackInfo ci) {
 
-        if (ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_QUICK_MOVE_THROW().getValue() &&
-                !LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() &&
-                (actionType == SlotActionType.QUICK_MOVE || (actionType == SlotActionType.THROW && button == 1))) {
-            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, button)) {
+        boolean move = actionType == SlotActionType.QUICK_MOVE;
+        boolean thr = actionType == SlotActionType.THROW;
+        if(!LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() && (move || thr)) {
+            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, thr, button)) {
                 ci.cancel();
             }
         }