+++ platforms/forge-1.17/src/main/java/org/anti_ad/mc/ipnext/mixin/MixinScreenHandler.java
@@ -4,7 +4,9 @@ package org.anti_ad.mc.ipnext.mixin;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.ClickType;
+import org.anti_ad.mc.ipnext.config.GuiSettings;
 import org.anti_ad.mc.ipnext.config.ModSettings;
+import org.anti_ad.mc.ipnext.event.ContinuousCraftingHandler;
 import org.anti_ad.mc.ipnext.event.LockSlotsHandler;
 import org.anti_ad.mc.ipnext.event.LockedSlotKeeper;
 import org.spongepowered.asm.mixin.Mixin;
@@ -27,14 +29,30 @@ public class MixinScreenHandler {
                                          ClickType actionType,
                                          Player player,
                                          CallbackInfo cir) {
-        if(ModSettings.INSTANCE.getLOCKED_SLOTS_DISABLE_QUICK_MOVE_THROW().getValue() &&
-                !LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() &&
-                (actionType == ClickType.QUICK_MOVE || (actionType == ClickType.THROW && button == 1))) {
-            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, button)) {
-              cir.cancel();
+        boolean move = actionType == ClickType.QUICK_MOVE;
+        boolean thr = actionType == ClickType.THROW;
+        if(!LockedSlotKeeper.INSTANCE.getProcessingLockedPickups() && (move || thr)) {
+            if (!LockSlotsHandler.INSTANCE.isQMoveActionAllowed(slotIndex, thr, button)) {
+                cir.cancel();
             }
+        }
+        if (GuiSettings.INSTANCE.getCONTINUOUS_CRAFTING_SAVED_VALUE().getValue()) {
+            ContinuousCraftingHandler.INSTANCE.setProcessingClick(true);
         }
     }
 
+    @Inject(at = @At(value = "TAIL",
+            target = "Lnet/minecraft/client/multiplayer/MultiPlayerGameMode;handleInventoryMouseClick(IIILnet/minecraft/world/inventory/ClickType;Lnet/minecraft/world/entity/player/Player;)V"),
+            method = "handleInventoryMouseClick")
+    public void postInternalOnSlotClickBegin(int windowId,
+                                             int slotIndex,
+                                             int button,
+                                             ClickType actionType,
+                                             Player player,
+                                             CallbackInfo cir) {
+        if (GuiSettings.INSTANCE.getCONTINUOUS_CRAFTING_SAVED_VALUE().getValue()) {
+            ContinuousCraftingHandler.INSTANCE.setProcessingClick(false);
+        }
+    }
 
 }